Moved AuthenticationClient::Impl to separate file.
To lighten the AuthenticationClient.cpp file and align pimpl to other
classes, this commit moves AuthenticationClient::Impl to a separate
file and renames it to AuthenticationClientImpl.

Relates-To: OLPEDGE-1161
